requirements.txt
openai>=1.0.0
python-pptx>=0.6.21